Added wall movement and death detection
Green walls move up and down. This code can be added to other objects and adjusted by the speed and height variables.
Also added death detection tot the player: When the player touches a wall it "dies" (for now only a debug message apears). The player can die with any object that has the tag death and a isTrigger collider